Some fixes

+++ test.py
@@ -8,6 +8,8 @@ import skimage.transform
 import numpy as np
 from keras.applications.inception_v3 import preprocess_input
 
+from models import extractor
+
 def predict_labels(video_path, extractor, img_shape, batch_size=64):
     reader = skvideo.io.FFmpegReader(video_path)
     num_frame = 0
@@ -16,7 +18,7 @@ def predict_labels(video_path, extractor, img_shape, batch_size=64):
     for frame in reader.nextFrame():
         frame = cv2.resize(frame, dsize=img_shape)
         frames.append(frame)
-    result = extractor.predict(preprocess_input(np.array(frames[:302])), batch_size=batch_size)
+    result = extractor.predict(preprocess_input(np.array(frames[:302]).astype(float)), batch_size=batch_size)
     return result
 
 def main():
